Revert "Fixed dashboard and support, page load optimizations, and directory structure reorganization"
