+++ localdb/monthlybudgets.csv
@@ -1,18 +1,17 @@
 budgetid,userid,totalbudget,budgetmonth
-1,1,2000,2023-01-01
-2,1,2000,2023-02-01
-3,1,2000,2023-03-01
-4,1,2000,2023-04-01
-5,1,2000,2023-05-01
-6,1,2000,2023-06-01
-7,1,2000,2023-07-01
-8,1,2000,2023-08-01
-9,1,2000,2023-09-01
-10,1,2000,2023-10-01
-11,1,2000,2023-11-01
-12,1,2000,2023-12-01
-13,1,2000,2024-06-01
-15,4,2100,2024-05-01
-16,4,1750,2024-04-01
-17,4,1500,2024-03-01
-14,4,2000,2024-06-01
+1,1,2000.0,2023-01-01
+2,1,2000.0,2023-02-01
+3,1,2000.0,2023-03-01
+4,1,2000.0,2023-04-01
+5,1,1250.0,2023-05-01
+6,1,2000.0,2023-06-01
+7,1,2000.0,2023-07-01
+8,1,2000.0,2023-08-01
+9,1,2000.0,2023-09-01
+10,1,2000.0,2023-10-01
+11,1,2000.0,2023-11-01
+12,1,2000.0,2023-12-01
+13,1,2000.0,2024-05-01
+14,1,300.0,2024-06-01
+15,4,1500.0,2024-01-01
+16,4,2500.0,2024-06-01
