--- project/callbacks/__init__.py
@@ -1,6 +1,0 @@
-from .dashboard_callback import dashboard_callback
-from .spendings_callback import spendings_callback
-from .sidebar_callback import sidebar_callback
-from .authentication_callback import authentication_callback
-from .settings_callback import settings_callback
-from .support_callback import support_callback
