--- project/config/gunicorn_config.py
@@ -1,4 +1,0 @@
-workers = 8
-threads = 4
-timeout = 120
-bind = "0.0.0.0:8000"
