--- project/config/wsgi.py
@@ -1,4 +1,0 @@
-from myfinanceplanner import application
-
-if __name__ == "__main__":
-    application.run()
