--- project/layouts/__init__.py
@@ -1,7 +1,0 @@
-from .dashboard_page import dashboard_page
-from .spendings_page import spendings_page
-from .welcome_page import welcome_page
-from .sign_in_page import sign_in_page
-from .sign_up_page import sign_up_page
-from .settings_page import settings_page
-from .support_page import support_page
