+++ project/requirements.txt
@@ -1,0 +1,5 @@
+dash==2.17.0
+pandas==2.0.2
+plotly==5.22.0
+SQLAlchemy==2.0.30
+psycopg2==2.9.9
